Court: allow to change config
Address PR #150 comments.
