+++ test/court/court-config.js
@@ -6,7 +6,7 @@ const { outcomeFor } = require('../helpers/crvoting')
 
 contract('Court config', ([_, sender, disputer, drafter, appealMaker, appealTaker, juror500, juror1000, juror3000]) => {
   let courtHelper, court
-  let initialConfig
+  let originalConfig, initialConfig
   let feeToken
   let jurorFee, heartbeatFee, draftFee, settleFee
   let commitTerms, revealTerms, appealTerms, appealConfirmTerms
@@ -40,6 +40,15 @@ contract('Court config', ([_, sender, disputer, drafter, appealMaker, appealTake
     appealCollateralFactor = bn(4)
     appealConfirmCollateralFactor = bn(6)
 
+    originalConfig = {
+      feeToken,
+      jurorFee, heartbeatFee, draftFee, settleFee,
+      commitTerms, revealTerms, appealTerms, appealConfirmTerms,
+      penaltyPct, finalRoundReduction,
+      firstRoundJurorsNumber, appealStepFactor, maxRegularAppealRounds,
+      appealCollateralFactor, appealConfirmCollateralFactor
+    }
+
     initialConfig = {
       newFeeTokenAddress: feeToken.address,
       newJurorFee: jurorFee,
@@ -70,107 +79,9 @@ contract('Court config', ([_, sender, disputer, drafter, appealMaker, appealTake
     })
   })
 
-  const checkConfig = async (termId, {
-    newFeeTokenAddress,
-    newJurorFee, newHeartbeatFee, newDraftFee, newSettleFee,
-    newCommitTerms, newRevealTerms, newAppealTerms, newAppealConfirmTerms,
-    newPenaltyPct, newFinalRoundReduction,
-    newFirstRoundJurorsNumber, newAppealStepFactor, newMaxRegularAppealRounds,
-    newAppealCollateralFactor, newAppealConfirmCollateralFactor
-  }) => {
-    const {
-      feeToken,
-      jurorFee, heartbeatFee, draftFee, settleFee,
-      commitTerms, revealTerms, appealTerms, appealConfirmTerms,
-      penaltyPct, finalRoundReduction,
-      firstRoundJurorsNumber, appealStepFactor, maxRegularAppealRounds,
-      appealCollateralFactor, appealConfirmCollateralFactor,
-    } = await courtHelper.getCourtConfig(termId)
-
-    assert.equal(feeToken, newFeeTokenAddress, 'Fee token does not match')
-    assertBn(jurorFee, newJurorFee, 'Juror fee does not match')
-    assertBn(heartbeatFee, newHeartbeatFee, 'Heartbeat fee does not match')
-    assertBn(draftFee, newDraftFee, 'Draft fee does not match')
-    assertBn(settleFee, newSettleFee, 'Settle fee does not match')
-    assertBn(commitTerms, newCommitTerms, 'Commit terms number does not match')
-    assertBn(revealTerms, newRevealTerms, 'Reveal terms number does not match')
-    assertBn(appealTerms, newAppealTerms, 'Appeal terms number does not match')
-    assertBn(appealConfirmTerms, newAppealConfirmTerms, 'Appeal confirmation terms number does not match')
-    assertBn(penaltyPct, newPenaltyPct, 'Penalty permyriad does not match')
-    assertBn(finalRoundReduction, newFinalRoundReduction, 'Final round reduction does not match')
-    assertBn(firstRoundJurorsNumber, newFirstRoundJurorsNumber, 'First round jurors number does not match')
-    assertBn(appealStepFactor, newAppealStepFactor, 'Appeal step factor does not match')
-    assertBn(maxRegularAppealRounds, newMaxRegularAppealRounds, 'Number af max regular appeal rounds does not match')
-    assertBn(appealCollateralFactor, newAppealCollateralFactor, 'Appeal collateral factor does not match')
-    assertBn(appealConfirmCollateralFactor, newAppealConfirmCollateralFactor, 'Appeal confirmation collateral factor does not match')
-  }
-
-  const buildNewConfig = async (iteration = 1) => {
-    const newFeeToken = await artifacts.require('ERC20Mock').new('Court Fee Token', 'CFT', 18)
-    const newFeeTokenAddress = newFeeToken.address
-    const newJurorFee = jurorFee.add(bigExp(iteration * 10, 18))
-    const newHeartbeatFee = heartbeatFee.add(bigExp(iteration * 10, 18))
-    const newDraftFee = draftFee.add(bigExp(iteration * 10, 18))
-    const newSettleFee = settleFee.add(bigExp(iteration * 10, 18))
-    const newCommitTerms = commitTerms.add(bn(iteration))
-    const newRevealTerms = revealTerms.add(bn(iteration))
-    const newAppealTerms = appealTerms.add(bn(iteration))
-    const newAppealConfirmTerms = appealConfirmTerms.add(bn(iteration))
-    const newPenaltyPct = penaltyPct.add(bn(iteration * 100))
-    const newFinalRoundReduction = finalRoundReduction.add(bn(iteration * 100))
-    const newFirstRoundJurorsNumber = firstRoundJurorsNumber.add(bn(iteration))
-    const newAppealStepFactor = appealStepFactor.add(bn(iteration))
-    const newMaxRegularAppealRounds = maxRegularAppealRounds.add(bn(iteration))
-    const newAppealCollateralFactor = appealCollateralFactor.add(bn(iteration))
-    const newAppealConfirmCollateralFactor = appealConfirmCollateralFactor.add(bn(iteration))
-
-    return {
-      newFeeTokenAddress,
-      newJurorFee, newHeartbeatFee, newDraftFee, newSettleFee,
-      newCommitTerms, newRevealTerms, newAppealTerms, newAppealConfirmTerms,
-      newPenaltyPct, newFinalRoundReduction,
-      newFirstRoundJurorsNumber, newAppealStepFactor, newMaxRegularAppealRounds,
-      newAppealCollateralFactor,
-      newAppealConfirmCollateralFactor,
-    }
-  }
-
-  const changeConfigPromise = async (termId, from, iteration = 1) => {
-    const newConfig = await buildNewConfig()
-    const {
-      newFeeTokenAddress,
-      newJurorFee, newHeartbeatFee, newDraftFee, newSettleFee,
-      newCommitTerms, newRevealTerms, newAppealTerms, newAppealConfirmTerms,
-      newPenaltyPct, newFinalRoundReduction,
-      newFirstRoundJurorsNumber, newAppealStepFactor, newMaxRegularAppealRounds,
-      newAppealCollateralFactor,
-      newAppealConfirmCollateralFactor,
-    } = newConfig
-
-    const promise = court.setCourtConfig(
-      termId,
-      newFeeTokenAddress,
-      [ newJurorFee, newHeartbeatFee, newDraftFee, newSettleFee ],
-      [ newCommitTerms, newRevealTerms, newAppealTerms, newAppealConfirmTerms ],
-      [ newPenaltyPct, newFinalRoundReduction ],
-      [ newFirstRoundJurorsNumber, newAppealStepFactor, newMaxRegularAppealRounds ],
-      [ newAppealCollateralFactor, newAppealConfirmCollateralFactor ],
-      { from }
-    )
-
-    return { promise, newConfig }
-  }
-
-  const changeConfig = async (termId, iteration = 1) => {
-    const { promise, newConfig } = await changeConfigPromise(termId, courtHelper.governor)
-    await promise
-
-    return newConfig
-  }
-
   context('initialization', () => {
     it('config is properly set', async () => {
-      await checkConfig(0, initialConfig)
+      await courtHelper.checkConfig(0, initialConfig)
     })
   })
 
@@ -187,7 +98,7 @@ contract('Court config', ([_, sender, disputer, drafter, appealMaker, appealTake
       // make sure account used is not governor
       assert.notEqual(courtHelper.governor, from, 'it is actually governor!')
 
-      const { promise } = await changeConfigPromise(configChangeTermId, sender)
+      const { promise } = await courtHelper.changeConfigPromise(originalConfig, configChangeTermId, sender)
       await assertRevert(promise, ERROR_BAD_SENDER)
     })
 
@@ -196,7 +107,7 @@ contract('Court config', ([_, sender, disputer, drafter, appealMaker, appealTake
       // move forward
       await courtHelper.setTerm(configChangeTermId + 1)
 
-      const { promise } = await changeConfigPromise(configChangeTermId, courtHelper.governor)
+      const { promise } = await courtHelper.changeConfigPromise(originalConfig, configChangeTermId, courtHelper.governor)
       await assertRevert(promise, ERROR_TOO_OLD_TERM)
     })
 
@@ -205,7 +116,7 @@ contract('Court config', ([_, sender, disputer, drafter, appealMaker, appealTake
       // move forward
       await courtHelper.setTerm(configChangeTermId - 1)
 
-      const { promise } = await changeConfigPromise(configChangeTermId, courtHelper.governor)
+      const { promise } = await courtHelper.changeConfigPromise(originalConfig, configChangeTermId, courtHelper.governor)
       await assertRevert(promise, ERROR_TOO_OLD_TERM)
     })
 
@@ -214,49 +125,80 @@ contract('Court config', ([_, sender, disputer, drafter, appealMaker, appealTake
       let newConfig
 
       beforeEach('ask for the change', async () => {
-        newConfig = await changeConfig(configChangeTermId)
+        newConfig = await courtHelper.changeConfig(originalConfig, configChangeTermId)
       })
 
       it('check it from the past', async () => {
-        await checkConfig(configChangeTermId, newConfig)
+        await courtHelper.checkConfig(configChangeTermId, newConfig)
       })
 
       it('check once the change term id has been reached', async () => {
         // move forward
         await courtHelper.setTerm(configChangeTermId)
 
-        await checkConfig(configChangeTermId, newConfig)
+        await courtHelper.checkConfig(configChangeTermId, newConfig)
       })
     })
 
-    context('overwriting changes', () => {
+    context('overwriting changes at a later term', () => {
       const configChangeTermId1 = 3, configChangeTermId2 = 4
       let newConfig1, newConfig2
 
       beforeEach('ask for the changes', async () => {
-        newConfig1 = await changeConfig(configChangeTermId1)
-        newConfig2 = await changeConfig(configChangeTermId2, 2)
+        newConfig1 = await courtHelper.changeConfig(originalConfig, configChangeTermId1)
+        newConfig2 = await courtHelper.changeConfig(originalConfig, configChangeTermId2, 2)
       })
 
       it('check it from the past', async () => {
-        await checkConfig(configChangeTermId1, initialConfig)
-        await checkConfig(configChangeTermId2, newConfig2)
+        await courtHelper.checkConfig(configChangeTermId1, initialConfig)
+        await courtHelper.checkConfig(configChangeTermId2, newConfig2)
       })
 
       it('check once the change term id for the first change has been reached', async () => {
         // move forward
         await courtHelper.setTerm(configChangeTermId1)
 
-        await checkConfig(configChangeTermId1, initialConfig)
-        await checkConfig(configChangeTermId2, newConfig2)
+        await courtHelper.checkConfig(configChangeTermId1, initialConfig)
+        await courtHelper.checkConfig(configChangeTermId2, newConfig2)
       })
 
       it('check once the change term id for the second change has been reached', async () => {
         // move forward
         await courtHelper.setTerm(configChangeTermId2)
 
-        await checkConfig(configChangeTermId1, initialConfig)
-        await checkConfig(configChangeTermId2, newConfig2)
+        await courtHelper.checkConfig(configChangeTermId1, initialConfig)
+        await courtHelper.checkConfig(configChangeTermId2, newConfig2)
+      })
+    })
+
+    context('overwriting changes at a prior term', () => {
+      const configChangeTermId1 = 4, configChangeTermId2 = 3
+      let newConfig1, newConfig2
+
+      beforeEach('ask for the changes', async () => {
+        newConfig1 = await courtHelper.changeConfig(originalConfig, configChangeTermId1)
+        newConfig2 = await courtHelper.changeConfig(originalConfig, configChangeTermId2, 2)
+      })
+
+      it('check it from the past', async () => {
+        await courtHelper.checkConfig(configChangeTermId1, newConfig2)
+        await courtHelper.checkConfig(configChangeTermId2, newConfig2)
+      })
+
+      it('check once the change term id for the first change has been reached', async () => {
+        // move forward
+        await courtHelper.setTerm(configChangeTermId1)
+
+        await courtHelper.checkConfig(configChangeTermId1, newConfig2)
+        await courtHelper.checkConfig(configChangeTermId2, newConfig2)
+      })
+
+      it('check once the change term id for the second change has been reached', async () => {
+        // move forward
+        await courtHelper.setTerm(configChangeTermId2)
+
+        await courtHelper.checkConfig(configChangeTermId1, newConfig2)
+        await courtHelper.checkConfig(configChangeTermId2, newConfig2)
       })
     })
   })
@@ -282,7 +224,7 @@ contract('Court config', ([_, sender, disputer, drafter, appealMaker, appealTake
       disputeId = await courtHelper.dispute({ draftTermId, disputer })
 
       // change config
-      await changeConfig(configChangeTermId)
+      await courtHelper.changeConfig(originalConfig, configChangeTermId)
 
       // move forward to dispute start
       await courtHelper.setTerm(draftTermId)
